+++ operator/apis/platform/v1alpha1/common_types.go
@@ -39,6 +39,10 @@ type DeploymentSpec struct {
 	// If you want this component to only run on specific nodes, you can configure a node selector here.
 	//+operator-sdk:csv:customresourcedefinitions:type=spec,displayName="Node Selector",order=101
 	NodeSelector map[string]string `json:"nodeSelector,omitempty"`
+
+	// If you want this component to only run on specific nodes, you can configure tolerations of tainted nodes.
+	//+operator-sdk:csv:customresourcedefinitions:type=spec,xDescriptors={"urn:alm:descriptor:com.tectonic.ui:tolerations"},order=102
+	Tolerations []*corev1.Toleration `json:"tolerations,omitempty"`
 }
 
 // StackRoxCondition defines a condition for a StackRox custom resource.